Декомпозирует компонент tab-list
